Modified registerAdvisor page to use templating.
I fixed some stuff up and renamed the template to the same name as the page which uses it. With the templating system as it is, it doesn't make much sense to have a template be a page on it's own. In the future, I'd like to revise our template folder position in the directory structure, because right now it is available to the public. It should probably be outside the public_html, like around where class/ is. That is a problem which Mr. Lupoli may have when tinkering with our project (if we're picked).